pyproject.toml
[tool.poetry]
name = "volctrl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "GPL"
readme = "README.md"
packages = [
  { include = "volctrl.py" },
]

[tool.poetry.scripts]
volctrl = 'volctrl:run'

[tool.poetry.dependencies]
python = "^3.11"
pyalsaaudio = "^0.11.0"
evdev = "^1.7.1"
gpiozero = "^2.0.1"
adafruit-circuitpython-neopixel-spi = "^1.0.9"
rpi-gpio = "^0.7.1"
adafruit-circuitpython-led-animation = "^2.9.1"
lgpio = "^0.2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
